Use `nvim-tree.lua` instead of `netrw`
Reason: the collapsing tree view in `netrw` is super annoying. I
couldn't find a way to fix this. Apparently, the tree view of `netrw` is
quite buggy.

Though, at work (i.e. working with a huge mono-repo) a file explorer
with a tree view is quite handy.

Note that currently no file icons are displayed.